requirements.txt
asgiref==3.4.1
astroid==2.7.2
Django==3.2.6
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
flake8==3.9.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
platformdirs==2.2.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.10.2
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
wrapt==1.12.1
